+++ NEWS.md
@@ -1,3 +1,6 @@
+# Development version
+* Update `fpr_kable` to deal with duplicated Table captions as per #68
+* Update fpr_photo_amalg_cv to include full calls to functions, importFrom and move to stand alone script. replace calls to %like% with str_detect. add to photo family
 * Added fpr_dbq_clip which clips a point layer by a polygon layer.
 * Added fpr_create_hydrograph which creates hydrographs of flow data for a given station
 * change name of `fpr_t_site_id` to `fpr_tidy_assign_site_id` to clarify use and avoid future conflicts if/when